Resolve duplication in Samples.
