+++ src/Fixie.Samples/Parameterized/CustomConvention.cs
@@ -24,18 +24,6 @@
                 .Add<InputAttributeParameterSource>();
         }
 
-        class CreateInstancePerClass : Lifecycle
-        {
-            public void Execute(Type testClass, Action<CaseAction> runCases)
-            {
-                var instance = Activator.CreateInstance(testClass);
-
-                runCases(@case => @case.Execute(instance));
-
-                (instance as IDisposable)?.Dispose();
-            }
-        }
-
         class InputAttributeParameterSource : ParameterSource
         {
             public IEnumerable<object[]> GetParameters(MethodInfo method)
