+++ src/Fixie.Samples/Shuffle/CustomConvention.cs
@@ -19,17 +19,5 @@
                 .Lifecycle<CreateInstancePerClass>()
                 .ShuffleCases(new Random(Seed));
         }
-
-        class CreateInstancePerClass : Lifecycle
-        {
-            public void Execute(Type testClass, Action<CaseAction> runCases)
-            {
-                var instance = Activator.CreateInstance(testClass);
-
-                runCases(@case => @case.Execute(instance));
-
-                (instance as IDisposable)?.Dispose();
-            }
-        }
     }
 }
