+++ src/Fixie.Samples/Skipped/CustomConvention.cs
@@ -27,17 +27,5 @@
 
         static bool SkipDueToMethodLevelSkipAttribute(Case @case)
             => @case.Method.Has<SkipAttribute>();
-
-        class CreateInstancePerClass : Lifecycle
-        {
-            public void Execute(Type testClass, Action<CaseAction> runCases)
-            {
-                var instance = Activator.CreateInstance(testClass);
-
-                runCases(@case => @case.Execute(instance));
-
-                (instance as IDisposable)?.Dispose();
-            }
-        }
     }
 }
